Move run_until is_target_state comparison functions into testutils
